[skip ci] clean CUDA deps
